capstone_stock_trading_policy_optimization
Stock investment can be one of the ways to manage one's asset. In this study we optimized hyper-parameters of abcd stock trading policy with reinforcement learning policy-based method and compared the return of investment (ROI) of our optimized trading policy with buy-and-hold trading policy. Results showed that our optimized policy outperforms buy-and-hold strategy in terms of ROI by 10.98\% and 5.37\% per annum on average for testing data for four-years for AAPL and SPY, respectively.  Furthermore, our optimized policy outperforms buy-and-hold strategy in terms of Sortino Ratio by  9.8\% and 7.2\% on average for testing data for four-years for AAPL and SPY, respectively. This shows that our optimized policy not only improves ROI, but also reduces downside risk. An iPython Notebook is created such that the trading policy can be used to generate buy, sell or do nothing suggestions for Yahoo Finance data by entering the stock ticker to the iPython Notebook.